create shared base group source and inherit from it
