+++ app/src/main/shared/sources/ModellingGroupSource.ts
@@ -1,13 +1,12 @@
 import { Source } from "./Source";
-import { ModellingGroup, ModellingGroupDetails } from "../models/Generated";
+import { ModellingGroup } from "../models/Generated";
 import SourceModel = AltJS.SourceModel;
 import { modellingGroupActions } from "../actions/ModellingGroupActions";
-import {RemoteContent} from "../models/RemoteContent";
-import {ILookup} from "../models/Lookup";
 
-export class ModellingGroupSource extends Source<GroupState> {
+
+export class GroupSourceBase<TState> extends Source<TState> {
+
     fetchGroups: () => SourceModel<ModellingGroup[]>;
-    fetchGroupDetails: () => SourceModel<ModellingGroupDetails>;
 
     constructor() {
         super();
@@ -15,16 +14,5 @@ export class ModellingGroupSource extends Source<GroupState> {
             success: modellingGroupActions.updateGroups,
             loading: modellingGroupActions.beginFetchGroups
         });
-        this.fetchGroupDetails = () => this.doFetch(s => `/modelling-groups/${s.currentGroupId}/`, {
-            success: modellingGroupActions.updateGroupDetails,
-            loading: modellingGroupActions.beginFetchDetails
-        });
     }
-}
-
-
-export interface GroupState extends RemoteContent {
-    groups: ModellingGroup[];
-    groupDetails: ILookup<ModellingGroupDetails>;
-    currentGroupId: string;
 }
